replace mysql with mongoDb
We have to format and parse addresses every time we store/read them from mysql. We could prevent this if we store them as json. Which is ugly.

A better way is to store the whole mail/attachment as document. The addresses in to,cc and bcc stay easily accessible that way